+++ templates/generate_done.html
@@ -15,7 +15,7 @@
       <li><a href="{{ parquet }}" class="link-light">Télécharger le jeu de données</a></li>
       <li><a href="{{ embeddings }}" class="link-light">Télécharger les embeddings</a></li>
     </ul>
-    <a href="{{ url_for('cluster_page') }}?uid={{ uid }}" class="btn btn-primary">Aller au clustering</a>
+    <a href="{{ url_for('cluster.cluster_page') }}?uid={{ uid }}" class="btn btn-primary">Aller au clustering</a>
   </div>
 </body>
 </html>
