+++ templates/home.html
@@ -12,8 +12,8 @@
     <h1 class="mb-4">Edu Topics</h1>
     <p class="lead">Choisissez une action</p>
     <div class="d-flex justify-content-center gap-3 mt-4">
-      <a href="{{ url_for('generate_page') }}" class="btn btn-primary btn-lg">Générer des embeddings</a>
-      <a href="{{ url_for('cluster_page') }}" class="btn btn-secondary btn-lg">Clustering</a>
+      <a href="{{ url_for('generate.generate_page') }}" class="btn btn-primary btn-lg">Générer des embeddings</a>
+      <a href="{{ url_for('cluster.cluster_page') }}" class="btn btn-secondary btn-lg">Clustering</a>
     </div>
   </div>
 </body>
