+++ templates/results.html
@@ -77,7 +77,7 @@
       <div class="d-flex gap-2">
         <a id="dl" class="btn btn-primary">Télécharger le Parquet</a>
         <a id="dlEmb" class="btn btn-secondary">Télécharger Embeddings</a>
-        <a href="{{ url_for('landing') }}" class="btn btn-outline-light">Nouvelle analyse</a>
+        <a href="{{ url_for('common.landing') }}" class="btn btn-outline-light">Nouvelle analyse</a>
       </div>
     </div>
